requirements.txt

transformers[sentencepiece]
torch
sentence-transformers
faiss-cpu
